# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: HRM + Distributor ERP Platform API — Phase 1 (Auth, Employees, Attendance, Leave)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful or MFA required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout — revoke refresh token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutInput"
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /v1/auth/change-password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent (always returns 200 to prevent enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/auth/mfa/setup:
    post:
      operationId: setupMfa
      tags: [auth]
      summary: Enable MFA — returns QR code and secret
      security:
        - bearerAuth: []
      responses:
        "200":
          description: MFA setup info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaSetupResponse"
  /v1/auth/mfa/verify:
    post:
      operationId: verifyMfa
      tags: [auth]
      summary: Verify MFA code and complete login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaVerifyInput"
      responses:
        "200":
          description: MFA verified — returns full tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
  /v1/auth/mfa/disable:
    post:
      operationId: disableMfa
      tags: [auth]
      summary: Disable MFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaDisableInput"
      responses:
        "200":
          description: MFA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ── Companies ──────────────────────────────────────────────────────────
...
  /v1/attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      security:
        - bearerAuth: []
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["string", "null"]
        - name: from
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: to
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceListResponse"
    post:
      operationId: logAttendance
      tags: [attendance]
      summary: Log attendance (check-in or check-out)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Attendance logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /v1/attendance/{id}:
    get:
      operationId: getAttendance
      tags: [attendance]
      summary: Get a single attendance record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ── Leave ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
      required: [success, message, timestamp]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: ["string", "null"]
          required: [message]
        timestamp:
          type: string
      required: [success, error, timestamp]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ── Auth Schemas ────────────────────────────────────────────────────────
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    TokenUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        companyId:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
      required: [id, email, role]
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/TokenUser"
      required: [accessToken, refreshToken, user]
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: ["string", "null"]
        refreshToken:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/TokenUser"
        requiresMfa:
          type: ["boolean", "null"]
        mfaToken:
          type: ["string", "null"]
      required: []
    LogoutInput:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    RefreshInput:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
      required: [accessToken]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        isEmailVerified:
          type: boolean
        mfaEnabled:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
        companyId:
          type: ["string", "null"]
        employeeId:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, role, isActive, isEmailVerified, mfaEnabled, createdAt]
    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
    PasswordChangeInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
      required: [email]
    ResetPasswordInput:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
      required: [token, newPassword]
    MfaSetupResponse:
      type: object
      properties:
        secret:
          type: string
        qrCodeUrl:
          type: string
        otpauthUrl:
          type: string
      required: [secret, qrCodeUrl, otpauthUrl]
    MfaVerifyInput:
      type: object
      properties:
        mfaToken:
          type: string
        code:
          type: string
      required: [mfaToken, code]
    MfaDisableInput:
      type: object
      properties:
        password:
          type: string
        code:
          type: string
      required: [password, code]

    # ── Company Schemas ─────────────────────────────────────────────────────
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        employeeName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        status:
          type: string
        hoursWorked:
          type: ["number", "null"]
        overtime:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, employeeId, date, status, source, createdAt, updatedAt]
    AttendanceListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        pagination:
          $ref: "#/components/schemas/Pagination"
        timestamp:
          type: string
      required: [success, data, pagination, timestamp]
    AttendanceInput:
      type: object
      properties:
        employeeId:
          type: string
        date:
          type: string
          format: date
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        status:
          type: string
          default: PRESENT
        notes:
          type: ["string", "null"]
        source:
          type: string
          default: MANUAL
      required: [employeeId, date]
    AttendanceUpdate:
      type: object
      properties:
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        hoursWorked:
          type: ["number", "null"]
        overtime:
          type: ["number", "null"]
...