# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HRM + Distributor ERP Platform API — Phase 1 (Auth, Employees, Attendance, Leave)
 * OpenAPI spec version: 1.0.0
 */

export interface PasswordChangeInput {
  currentPassword: string;
  newPassword: string;
}